Add light control
Former-commit-id: 6def92495fab00c93ba61f1379befc7fcc1d8df6

+++ renderer/core/platform.c
@@ -116,10 +116,10 @@ void private_blit_buffer_rgb(colorbuffer_t *src, image_t *dst) {
 
 double private_get_raw_time(void);
 
-double input_get_time(void) {
+float input_get_time(void) {
     static double initial = -1;
     if (initial < 0) {
         initial = private_get_raw_time();
     }
-    return private_get_raw_time() - initial;
+    return (float)(private_get_raw_time() - initial);
 }
